requirements.txt
# Core Data Science
numpy>=1.21.0
pandas>=1.3.0
matplotlib>=3.4.0
seaborn>=0.11.0

# Image Processing
Pillow>=9.0.0
opencv-python>=4.5.0
pillow-heif>=0.7.0

# Machine Learning
torch>=2.0.0
torchvision>=0.15.0
scikit-learn>=1.0.0

# Utilities
tqdm>=4.62.0
pyyaml>=6.0

# Jupyter (for notebooks)
jupyter>=1.0.0
ipykernel>=6.0.0

# Optional but useful
tensorboard>=2.10.0
